remove transformer_decoder_layer_mp.py
